requirements.txt
# Trading Engine Requirements

# Core
redis>=4.5.0
psycopg2-binary>=2.9.0
requests>=2.28.0

# WebSocket for data feeds
websocket-client>=1.5.0

# API Server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Strategy hot-reload
watchdog>=3.0.0

# Screenshot capture (optional)
selenium>=4.15.0
mss>=9.0.0

# Utilities
python-dateutil>=2.8.0
